requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
fastapi==0.54.0
feedparser==5.2.1
h11==0.9.0
httptools==0.1.1
idna==2.9
psycopg2-binary==2.8.4
pydantic==1.4
python-dotenv==0.12.0
requests==2.23.0
SQLAlchemy==1.3.13
starlette==0.13.2
urllib3==1.25.8
uvicorn==0.11.7
uvloop==0.14.0
websockets==8.1
